requirements.txt
requests==2.23.0
pycryptodome==3.9.7
numpy==1.18.4
opencv-python==4.2.0.34
pytesseract==0.3.4
xlwt==1.3.0
xlrd==1.2.0
xlutils==2.0.0
